platform-api/requirements.txt
fastapi[all]>=0.104.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.20
asyncpg>=0.28.0
alembic>=1.12.0
pydantic-settings>=2.0.3
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
pytest>=7.4.2
pytest-asyncio>=0.21.1
ruff>=0.1.3
black>=23.9.1
mypy>=1.6.0